Fix a bug when the slider reached its maximum value.

+++ internal/server/writers/ticker/timeline/timeline.go
@@ -20,7 +20,7 @@ import (
 type Timeline struct {
 	mux sync.Mutex
 
-	currentTick uint64
+	currentTick int64
 	displayTick int64
 	db          *timedb.TimeDB
 }
@@ -49,18 +49,18 @@ func (tl *Timeline) MarshalDisplay() *pb.TimeLine {
 		float64(maxStorage),
 		int(float64(currentStorage)/float64(maxStorage)*100))
 	return &pb.TimeLine{
-		DisplayTick:     uint64(tl.adjustDisplayTick(tl.displayTick)),
+		DisplayTick:     tl.adjustDisplayTick(tl.displayTick),
 		OldestTick:      tl.db.Oldest(),
-		HistoryLength:   uint64(tl.db.NumRecords()),
+		HistoryLength:   int64(tl.db.NumRecords()),
 		StorageCapacity: storageCapacity,
 	}
 }
 
-func (tl *Timeline) setDisplayTick(displayTick uint64) {
+func (tl *Timeline) setDisplayTick(displayTick int64) {
 	tl.mux.Lock()
 	defer tl.mux.Unlock()
 
-	tl.displayTick = int64(displayTick)
+	tl.displayTick = displayTick
 }
 
 func (tl *Timeline) offsetDisplayTick(offset int64) {
@@ -102,12 +102,12 @@ func (tl *Timeline) adjustDisplayTick(tick int64) (out int64) {
 		}
 	}()
 
-	oldest := int64(tl.db.Oldest())
-	if tick < int64(tl.db.Oldest()) {
+	oldest := tl.db.Oldest()
+	if tick < tl.db.Oldest() {
 		// Adjust to the oldest tick.
 		return oldest
 	}
-	if tick < int64(tl.currentTick) {
+	if tick < tl.currentTick {
 		// Valid value between the oldest tick and the current tick.
 		return tick
 	}
@@ -116,7 +116,7 @@ func (tl *Timeline) adjustDisplayTick(tick int64) (out int64) {
 		return 0
 	}
 	// Adjust to the latest tick.
-	return int64(tl.currentTick) - 1
+	return tl.currentTick - 1
 }
 
 // MarshalData serializes the data given the current tick being displayed.
@@ -124,7 +124,7 @@ func (tl *Timeline) MarshalData(data *treepb.NodeData, path []string) {
 	tl.mux.Lock()
 	defer tl.mux.Unlock()
 
-	displayTick := uint64(tl.adjustDisplayTick(tl.displayTick))
+	displayTick := tl.adjustDisplayTick(tl.displayTick)
 	rec := tl.db.Fetch(displayTick)
 	if rec == nil {
 		data.Error = fmt.Sprintf("data for tick %d does not exist", displayTick)
@@ -142,7 +142,7 @@ func (tl *Timeline) MarshalData(data *treepb.NodeData, path []string) {
 }
 
 // CurrentTick returns the next frame to store.
-func (tl *Timeline) CurrentTick() uint64 {
+func (tl *Timeline) CurrentTick() int64 {
 	tl.mux.Lock()
 	defer tl.mux.Unlock()
 
@@ -154,7 +154,7 @@ func (tl *Timeline) IsLastTickDisplayed() bool {
 	tl.mux.Lock()
 	defer tl.mux.Unlock()
 
-	return tl.displayTick >= int64(tl.currentTick)
+	return tl.displayTick >= tl.currentTick
 }
 
 // Close the timeline and release all the memory.
